ESWE-1735: Refactor tests
- Reduce contexts (read-only mode on or off)
- Move `TestBase` to `integration` package
- Suppress warnings
- Remove Unnecessary Spring context (from `ClientTrackingWebFilterTest`)
