+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsresettlementpassportapi/config/ClientTrackingWebFilterTest.kt
@@ -4,34 +4,17 @@ import io.opentelemetry.api.common.AttributeKey
 import io.opentelemetry.api.trace.Tracer
 import io.opentelemetry.sdk.testing.junit5.OpenTelemetryExtension
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.junit.jupiter.api.extension.RegisterExtension
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.ConfigDataApplicationContextInitializer
-import org.springframework.context.annotation.Import
 import org.springframework.http.HttpHeaders
 import org.springframework.mock.http.server.reactive.MockServerHttpRequest
 import org.springframework.mock.web.server.MockServerWebExchange
-import org.springframework.test.context.ActiveProfiles
-import org.springframework.test.context.ContextConfiguration
-import org.springframework.test.context.junit.jupiter.SpringExtension
 import org.springframework.web.server.WebFilterChain
 import reactor.core.publisher.Mono
 import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.helpers.JwtAuthHelper
 
-@Import(JwtAuthHelper::class, ClientTrackingWebFilter::class)
-@ContextConfiguration(initializers = [ConfigDataApplicationContextInitializer::class])
-@ActiveProfiles("test")
-@ExtendWith(SpringExtension::class)
 class ClientTrackingWebFilterTest {
-
-  @Suppress("SpringJavaInjectionPointsAutowiringInspection")
-  @Autowired
-  private lateinit var clientTrackingWebFilter: ClientTrackingWebFilter
-
-  @Suppress("SpringJavaInjectionPointsAutowiringInspection")
-  @Autowired
-  private lateinit var jwtAuthHelper: JwtAuthHelper
+  private val clientTrackingWebFilter = ClientTrackingWebFilter()
+  private val jwtAuthHelper = JwtAuthHelper()
 
   private val tracer: Tracer = otelTesting.openTelemetry.getTracer("test")
   private val filterChain = WebFilterChain { Mono.empty() }
