+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsresettlementpassportapi/integration/CaseAllocationIntegrationReadOnlyModeTest.kt
@@ -1,27 +1,13 @@
 package uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.integration
 
 import io.mockk.every
-import io.mockk.mockk
 import io.mockk.mockkStatic
 import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.TestConfiguration
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Import
-import org.springframework.context.annotation.Primary
 import org.springframework.test.context.jdbc.Sql
 import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.data.CaseAllocation
-import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.service.FeatureFlagValueProvider
 import java.time.LocalDateTime
 
-@TestConfiguration
-class CaseAllocationTestMockConfig {
-  @Bean
-  @Primary
-  fun featureFlagValueProvider(): FeatureFlagValueProvider = mockk { every { isReadOnlyMode() } returns true }
-}
-
-@Import(CaseAllocationTestMockConfig::class)
-class CaseAllocationIntegrationReadOnlyModeTest : IntegrationTestBase() {
+class CaseAllocationIntegrationReadOnlyModeTest : ReadOnlyIntegrationTestBase() {
 
   private val fakeNow = LocalDateTime.parse("2023-08-17T12:00:01")
