+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsresettlementpassportapi/integration/ProfileResetIntegrationReadOnlyModeTest.kt
@@ -1,26 +1,11 @@
 package uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.integration
 
-import io.mockk.every
-import io.mockk.mockk
 import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.TestConfiguration
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Import
-import org.springframework.context.annotation.Primary
 import org.springframework.test.context.jdbc.Sql
 import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.data.ProfileReset
 import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.data.ResetReason
-import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.service.FeatureFlagValueProvider
 
-@TestConfiguration
-class ProfileResetTestMockConfig {
-  @Bean
-  @Primary
-  fun featureFlagValueProvider(): FeatureFlagValueProvider = mockk { every { isReadOnlyMode() } returns true }
-}
-
-@Import(ProfileResetTestMockConfig::class)
-class ProfileResetIntegrationReadOnlyModeTest : IntegrationTestBase() {
+class ProfileResetIntegrationReadOnlyModeTest : ReadOnlyIntegrationTestBase() {
 
   @Test
   @Sql("classpath:testdata/sql/seed-profile-reset.sql")
