+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsresettlementpassportapi/jpa/repository/RepositoryTestBase.kt
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.jpa.repository
 import jakarta.transaction.Transactional
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.jdbc.Sql
-import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.helpers.TestBase
+import uk.gov.justice.digital.hmpps.hmppsresettlementpassportapi.integration.TestBase
 
 @SpringBootTest
 @Transactional
